Patch 1

+++ expenses/views.py
@@ -3,6 +3,9 @@ from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
+from django.utils import timezone
+from django.core.files import File
+from django.db.models import Sum
 
 from django.http import HttpResponseRedirect
 from django.urls import reverse
@@ -21,6 +24,78 @@ import time
 import requests
 import json
 import datetime
+
+from django.template.loader import get_template
+from django.http import HttpResponse
+from django.shortcuts import redirect
+from django.urls import reverse
+from io import BytesIO
+from reportlab.pdfgen import canvas
+from reportlab.lib.pagesizes import letter
+from reportlab.lib import colors
+from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
+from reportlab.platypus import SimpleDocTemplate, Paragraph, Table, TableStyle
+
+def generate_invoice_pdf(invoice_data, output_filename='invoice.pdf'):
+    buffer = BytesIO()
+    doc = SimpleDocTemplate(buffer, pagesize=letter)
+
+    elements = []
+
+    title_style = getSampleStyleSheet()['Title']
+    elements.append(Paragraph("Report Generated", title_style))
+
+    invoice_style = ParagraphStyle('InvoiceStyle', parent=getSampleStyleSheet()['BodyText'], spaceAfter=15)
+    elements.append(Paragraph("Report Details:", invoice_style))
+
+    table_data = [['Expenses Name', 'Category', 'Date', 'Total']]
+    for item in invoice_data:
+        total_amount = float(item.total_amount)
+        total_amountstr = "{:,.2f}".format(total_amount)
+        table_data.append([item.expense_name, item.category, item.date(), total_amountstr])
+
+    
+
+    table_style = TableStyle([('BACKGROUND', (0, 0), (-1, 0), colors.grey),
+                              ('TEXTCOLOR', (0, 0), (-1, 0), colors.whitesmoke),
+                              ('ALIGN', (0, 0), (-1, -1), 'CENTER'),
+                              ('FONTNAME', (0, 0), (-1, 0), 'Helvetica-Bold'),
+                              ('BOTTOMPADDING', (0, 0), (-1, 0), 12),
+                              ('BACKGROUND', (0, 1), (-1, -1), colors.beige),
+                              ('GRID', (0, 0), (-1, -1), 1, colors.black)])
+    table = Table(table_data, style=table_style)
+    elements.append(table)
+
+    total = sum(item.total_amount for item in invoice_data)
+    total_amount = float(total)
+    total_amountstr = "\u20B1 {:,.2f}".format(total_amount)
+    elements.append(Paragraph(f"Total: {total_amountstr}", invoice_style))
+
+    doc.build(elements)
+    pdf = buffer.getvalue()
+    buffer.close()
+
+    return pdf
+
+def create_rand_id():
+        from django.utils.crypto import get_random_string
+        return get_random_string(length=13, 
+            allowed_chars='ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890')
+
+
+from roboflow import Roboflow
+rf = Roboflow(api_key="He28uTlpoqBXsFA1hHxS")
+project = rf.workspace().project("receipt-logo-detection")
+model = project.version(3).model
+
+# infer on a local image 
+# print(model.predict("your_image.jpg", confidence=40, overlap=30).json())
+
+# visualize your prediction
+# model.predict("your_image.jpg", confidence=40, overlap=30).save("prediction.jpg")
+
+# infer on an image hosted elsewhere
+# print(model.predict("URL_OF_YOUR_IMAGE", hosted=True, confidence=40, overlap=30).json())
 
 # AUTH
 def Register(request):
@@ -74,25 +149,33 @@ def Logout(request):
 @login_required(login_url='login')
 def HomePage(request):
     today = datetime.datetime.now()
+    today2 = datetime.datetime.today()
+    day = today2.day
+    month = today2.month
+    year = today2.year
 
-    uploads = Uploaded_Image_Expenses.objects.filter(user=request.user)
-    catergory1 = ExpensesCategory.objects.all()
-    expenses = Expenses.objects.filter(user=request.user)
     total_per_month = 0
     total_expenses = 0
     total_expenses_per_month = [0,0,0,0,0,0,0,0,0,0,0,0]
-    catergory = []
+    catergory = []    
+    total_data_last_seven_days = []
+    date_last_seven_days = []
+
+    uploads = Expenses.objects.filter(user=request.user)
+    total_today = Expenses.objects.filter(user=request.user, date_added__year=year, date_added__month=month, date_added__day=day)
+
+    expenses = Expenses.objects.filter(user=request.user, date_added__year=year)
+
+    # Category
+    catergory1 = ExpensesCategory.objects.all()
     catergory_expenses = [0] * len(catergory1)
-    print(catergory_expenses)
-    counter = 0
     
+    counter = 0
     for i in catergory1:
         catergory.append(i.category)
         categ_expenses = Expenses.objects.filter(user=request.user, category=i)
         for j in categ_expenses:
-            print(counter)
             if j.category == i.category:
-                print(categ_expenses)
                 catergory_expenses[counter] += j.total_amount
         counter = counter + 1
 
@@ -102,9 +185,29 @@ def HomePage(request):
         for j in range(0,13): 
             if j == i.date_added.month:
                 total_expenses_per_month[j-1] += i.total_amount
+
         total_expenses += i.total_amount
         if today.month == i.date_added.month:
             total_per_month += i.total_amount
+
+    last_7_days = today2-datetime.timedelta(days=7)
+
+    last_seven_days = Expenses.objects.filter(user=request.user, date_added__gte=last_7_days).\
+    extra({'day':"date_added"}).\
+    values('day').annotate(count=Sum('total_amount'))
+
+    for i in last_seven_days:
+        date = datetime.datetime.fromisoformat(str(i["day"])[:-6])
+        formatted_date = date.strftime("%B %d, %Y")
+        total_data_last_seven_days.append(i["count"])
+        date_last_seven_days.append(formatted_date)
+
+    tolal_today_amount = 0
+
+    for amount in total_today:
+        tolal_today_amount += amount.total_amount
+
+    
     
     total_amount = float(total_per_month)
     total_per_month = "₱ {:,.2f}".format(total_amount)
@@ -112,17 +215,19 @@ def HomePage(request):
     total_amount = float(total_expenses)
     total_expenses = "₱ {:,.2f}".format(total_amount)
     
-    print(total_per_month)
-    print(total_expenses)
-    print(catergory)
-    print(catergory_expenses)
+    total_today = float(tolal_today_amount)
+    total_today = "₱ {:,.2f}".format(total_today)
+
     context = {'total_per_month': total_per_month,
                'total_expenses': total_expenses,
                'total_expenses_per_month': total_expenses_per_month,
                'uploads': len(uploads),
                'catergory': len(catergory),
                'chart_category': catergory,
-               'chart_category_epenses':catergory_expenses}
+               'chart_category_epenses':catergory_expenses,
+               'total_data_last_seven_days': total_data_last_seven_days,
+               'date_last_seven_days': date_last_seven_days,
+               'total_today': total_today }
             
     return render(request, "expenses/homepage.html", context)
 
@@ -135,6 +240,7 @@ def Upload_Image(request):
     imagereceipt = ImageReceiptForm()
     if request.method == 'POST':
         receiptForm = ImageReceiptForm(request.POST, request.FILES)
+        # print(timezone.now)
         if receiptForm.is_valid():
             receiptForm.save(commit=False).user = request.user
             image = receiptForm.save()
@@ -146,19 +252,27 @@ def Upload_Image(request):
             })
             headers = {
             'Content-Type': 'application/json',
-            'Authorization': 'ApiKey jzbsoriano@iskolarngbayan.pup.edu.ph:7871a591-62c6-4817-90e8-f66a061087bd'
+            'Authorization': 'ApiKey joaquinzaki21@gmail.com:c3e7596b-6f0f-456e-b8eb-d96d1dcc553c'
             }
 
             response = requests.request("POST", url, headers=headers, data=payload, timeout=10000)
 
             result = json.loads(response.text)
-            
+
+            model.predict(image.image.url, hosted=True, confidence=40, overlap=30).save("expenses\prediction\prediction.jpg")
+            prediction_result = model.predict(image.image.url, hosted=True, confidence=40, overlap=30).json()
+
+            if prediction_result:
+                image.image.save('prediction.jpg', File(open('expenses\prediction\prediction.jpg', 'rb')))
+
             print(result)
+            print(prediction_result)
 
             keywords = Keywords.objects.all()
             if len(result):
                 for i in keywords:
-                    if i.keywords.lower() in result[0]['ocr'].lower():
+                    if i.keywords.lower() in prediction_result["predictions"][0]["class"].lower() or \
+                        i.keywords.lower() in result[0]['ocr'].lower():
                         print('{0} found'.format(i))
                         if 'total' in result[0]['fields']:
                             if result[0]['fields']['total']:
@@ -179,6 +293,8 @@ def Upload_Image(request):
                             print(type(result[0]['fields']['heightImperial']['value']))
                             total_amount = result[0]['fields']['heightImperial']['value']
                             total_amount = total_amount.replace(',', '')
+                            if total_amount == '':
+                                total_amount = 0
                             Expenses.objects.create(
                                 user=request.user, 
                                 expense_name=i.keywords.title(),
@@ -196,16 +312,18 @@ def Upload_Image(request):
                                 category=i.category
                             )
                             return redirect('upload_confirmation', pk=image.reference_number)
-                messages.info(request, "Receipt Can't Read by the OCR")
+                messages.info(request, "Invalid Reciept")
+                return redirect('upload_reciept')
             else:
-                Uploaded_Image_Expenses.objects.get(
-                    reference_number=image.reference_number
-                ).delete()
+                # Uploaded_Image_Expenses.objects.get(
+                #     reference_number=image.reference_number
+                # ).delete()
                 system_messages = messages.get_messages(request)
                 for message in system_messages:
                     # This iteration is necessary
                     pass
-                messages.info(request, "Receipt Can't Read by the OCR")
+                messages.info(request, "Invalid Reciept")
+                return redirect('upload_reciept')
     context = {'form': imagereceipt}
     return render(request, "expenses/upload_image.html", context)
 
@@ -224,8 +342,8 @@ def Upload_Image_Confirmation(request, pk):
     print(stockform)
     if request.method == 'POST':
         stockform = ExpenseForm(request.POST, request.FILES, instance=expenses)
+        print(stockform)
         if stockform.is_valid():
-            print(stockform)
             stockform.save(commit=False).user = request.user
             stockform.save()
             return redirect('expenses')
@@ -246,13 +364,14 @@ def Cancel(request, pk):
 
 @login_required(login_url='login')
 def Delete(request, pk):
-    upload = Uploaded_Image_Expenses.objects.get(
+    upload = Uploaded_Image_Expenses.objects.filter(
         reference_number=pk
     )
     expenses = Expenses.objects.get(
         rndid=pk
     )
-    upload.delete()
+    if upload:
+        upload.delete()
     expenses.delete()
     return redirect('expenses')
 
@@ -260,7 +379,7 @@ def Delete(request, pk):
 def Update(request, pk):
     catergory = ExpensesCategory.objects.all()
     expenses = Expenses.objects.get(rndid=pk)
-    receipt = Uploaded_Image_Expenses.objects.get(
+    receipt = Uploaded_Image_Expenses.objects.filter(
             reference_number=pk
         )
     stockform = ExpenseForm(instance=expenses)
@@ -272,17 +391,23 @@ def Update(request, pk):
             stockform.save(commit=False).user = request.user
             stockform.save()
             return redirect('expenses')
-    context = {"expense": expenses, "image": receipt, "category": catergory}
+    if receipt:
+        context = {"expense": expenses, "image": receipt[0], "category": catergory}
+    else:
+        context = {"expense": expenses, "category": catergory}
     return render(request, "expenses/update_upload.html", context)
 
 @login_required(login_url='login')
 def View(request, pk):
     catergory = ExpensesCategory.objects.all()
     expenses = Expenses.objects.get(rndid=pk)
-    receipt = Uploaded_Image_Expenses.objects.get(
+    receipt = Uploaded_Image_Expenses.objects.filter(
             reference_number=pk
         )
-    context = {"expense": expenses, "image": receipt, "category": catergory}
+    if receipt:
+        context = {"expense": expenses, "image": receipt[0], "category": catergory}
+    else:
+        context = {"expense": expenses, "category": catergory}
     return render(request, "expenses/view_upload.html", context)
 
 @login_required(login_url='login')
@@ -299,3 +424,65 @@ def ManageUpload(request):
 
     context = {'expenses': expenses, "category": category}
     return render(request, "expenses/expenses.html", context)
+
+@login_required(login_url='login')
+def AddExpenses(request):
+    catergory = ExpensesCategory.objects.all()
+    print(create_rand_id())
+    if request.method == 'POST':
+        stockform = AddExpenseForm(request.POST)
+        print(stockform)
+        if stockform.is_valid():
+            print(stockform)
+            stockform.save(commit=False).user = request.user
+            rand_id = create_rand_id()
+            stockform.save(commit=False).rndid = str(rand_id)
+            stockform.save()
+            return redirect('expenses')
+    context = {"category": catergory}
+    return render(request, "expenses/manual_add.html", context)
+
+@login_required(login_url='login')
+def Report(request):
+    if request.method == 'POST':
+        report_name = request.POST.get("expense_name")
+        filter_date = request.POST.get("filter")
+        date_added = request.POST.get("date_added")
+        dateobject = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().day
+        print(dateobject)
+        print(filter_date)
+        print(report_name)
+
+        if filter_date == "Year":
+            year = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().year
+
+            uploads = Expenses.objects.filter(user=request.user, date_added__year=year)
+        elif filter_date == "Month":
+            year = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().year
+            month = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().month
+
+            uploads = Expenses.objects.filter(user=request.user, date_added__year=year, date_added__month=month)
+        
+        else:
+            year = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().year
+            month = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().month
+            day = datetime.datetime.strptime(date_added, '%Y-%m-%d').date().day
+
+            uploads = Expenses.objects.filter(user=request.user, date_added__year=year, date_added__month=month, date_added__day=day)
+
+        pdf = generate_invoice_pdf(uploads)
+
+        # Create HTTP response with PDF content
+        response = HttpResponse(pdf, content_type='application/pdf')
+        response['Content-Disposition'] = 'attachment; filename="{}.pdf"'.format(report_name)
+        # Render a template with JavaScript for redirection
+        template = get_template('expenses/redirect_template.html')
+        context = {'redirect_url': reverse('expenses')}
+        response.write(template.render(context, request))
+
+        return response
+    
+    context = {}
+    return render(request, "expenses/generate_report.html", context)
+
+
